# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrainFlow API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type UserDtoRole = typeof UserDtoRole[keyof typeof UserDtoRole];


export const UserDtoRole = {
  student: 'student',
  trainer: 'trainer',
} as const;

export interface UserDto {
  id: string;
  name: string;
  email: string;
  role: UserDtoRole;
  onboardingDone?: string;
}

export interface AuthResponse {
  user: UserDto;
  isNewUser: boolean;
}

export interface StudentLoginRequest {
  email: string;
}

export interface TrainerLoginRequest {
  email: string;
  pin: string;
}

export interface OnboardingRequest {
  userId: string;
  name?: string;
  goal?: string;
  weightKg?: string;
}

export interface ExerciseDto {
  id: string;
  name: string;
  muscleGroup: string;
  equipment: string;
  description?: string;
  isGlobal?: boolean;
  isCustom?: boolean;
}

export interface CreateExerciseRequest {
  name: string;
  muscleGroup: string;
  equipment: string;
  description?: string;
}

export interface StudentDto {
  id: string;
  name: string;
  email: string;
  status: string;
  lastSession?: string;
  activePlanName?: string;
  goal?: string;
  weightKg?: string;
  heightCm?: string;
}

export interface InviteStudentRequest {
  name: string;
  email: string;
}

export interface WorkoutPlanDto {
  id: string;
  name: string;
  trainerId: string;
  studentId?: string;
  isPublished?: boolean;
}

export interface WorkoutDayDto {
  id: string;
  planId: string;
  dayName: string;
  focus: string;
  orderIndex?: number;
}

export interface PlanExerciseDto {
  id: string;
  dayId: string;
  exerciseId: string;
  sets: number;
  reps: string;
  restSeconds: number;
  notes?: string;
  orderIndex?: number;
}

export interface PlanExerciseFullDto {
  id: string;
  dayId: string;
  exerciseId: string;
  sets: number;
  reps: string;
  restSeconds: number;
  notes?: string;
  orderIndex?: number;
  exercise: ExerciseDto;
}

export interface WorkoutDayFullDto {
  id: string;
  planId: string;
  dayName: string;
  focus: string;
  orderIndex?: number;
  exercises: PlanExerciseFullDto[];
}

export interface WorkoutPlanFullDto {
  id: string;
  name: string;
  trainerId: string;
  studentId?: string;
  isPublished?: boolean;
  days: WorkoutDayFullDto[];
}

export interface CreatePlanRequest {
  name: string;
  studentId?: string;
  isPublished?: boolean;
}

export interface UpdatePlanRequest {
  name?: string;
  studentId?: string;
  isPublished?: boolean;
}

export interface CreateDayRequest {
  dayName: string;
  focus: string;
  orderIndex?: number;
}

export interface CreatePlanExerciseRequest {
  exerciseId: string;
  sets: number;
  reps: string;
  restSeconds: number;
  notes?: string;
  orderIndex?: number;
}

export interface LoggedSetDto {
  exerciseId: string;
  setIndex: number;
  reps: number;
  weightKg: number;
}

export interface WorkoutSessionDto {
  id: string;
  studentId: string;
  planId?: string;
  dayId?: string;
  startTime: number;
  endTime?: number;
  exerciseFocus?: string;
  loggedSets: LoggedSetDto[];
}

export interface CreateSessionRequest {
  studentId: string;
  planId?: string;
  dayId?: string;
  startTime: number;
  endTime?: number;
  exerciseFocus?: string;
  loggedSets: LoggedSetDto[];
}

export type ConversationDtoLastMessage = {
  id: string;
  senderId: string;
  text: string;
  timestamp: number;
};

export interface ConversationDto {
  id: string;
  isGroup: boolean;
  participantIds: string[];
  participantNames: string[];
  lastMessage?: ConversationDtoLastMessage;
}

export interface MessageDto {
  id: string;
  conversationId: string;
  senderId: string;
  senderName: string;
  text: string;
  timestamp: number;
}

export interface SendMessageRequest {
  senderId: string;
  senderName: string;
  text: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type ListExercisesParams = {
muscleGroup?: string;
search?: string;
};

export type ListPlansParams = {
studentId?: string;
};

export type ListSessionsParams = {
studentId: string;
};

export type ListConversationsParams = {
userId: string;
};