# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrainFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  # ── Simple mobile app auth ─────────────────────────────────────────────────
  /auth/student-login:
    post:
      operationId: studentLogin
      tags: [auth]
      summary: Student magic link login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentLoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/trainer-login:
    post:
      operationId: trainerLogin
      tags: [auth]
      summary: Trainer PIN login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainerLoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  # ── Domain routes ──────────────────────────────────────────────────────────
  /exercises:
    get:
      operationId: listExercises
      tags: [exercises]
      summary: List exercises
      parameters:
        - name: muscleGroup
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExerciseDto"
    post:
      operationId: createExercise
      tags: [exercises]
      summary: Create custom exercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExerciseRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseDto"
...
  /sessions:
    get:
      operationId: listSessions
      tags: [sessions]
      summary: List sessions for the current student
      parameters:
        - name: studentId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutSessionDto"
    post:
      operationId: createSession
      tags: [sessions]
      summary: Start or save a workout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutSessionDto"
...
  /me/onboarding:
    post:
      operationId: completeOnboarding
      tags: [users]
      summary: Save student onboarding data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Simple app auth schemas ──────────────────────────────────────────────
    UserDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, trainer]
        onboardingDone:
          type: string
      required: [id, name, email, role]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserDto"
        isNewUser:
          type: boolean
      required: [user, isNewUser]
    StudentLoginRequest:
      type: object
      properties:
        email:
          type: string
      required: [email]
    TrainerLoginRequest:
      type: object
      properties:
        email:
          type: string
        pin:
          type: string
      required: [email, pin]

    # ── Domain schemas ───────────────────────────────────────────────────────
    OnboardingRequest:
      type: object
      properties:
        userId:
          type: string
        name:
          type: string
        goal:
          type: string
        weightKg:
          type: string
      required: [userId]
    ExerciseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        muscleGroup:
          type: string
        equipment:
          type: string
        description:
          type: string
        isGlobal:
          type: boolean
        isCustom:
          type: boolean
      required: [id, name, muscleGroup, equipment]
    CreateExerciseRequest:
      type: object
      properties:
        name:
          type: string
        muscleGroup:
          type: string
        equipment:
          type: string
        description:
          type: string
      required: [name, muscleGroup, equipment]
...
    WorkoutSessionDto:
      type: object
      properties:
        id:
          type: string
        studentId:
          type: string
        planId:
          type: string
        dayId:
          type: string
        startTime:
          type: integer
        endTime:
          type: integer
        exerciseFocus:
          type: string
        loggedSets:
          type: array
          items:
            $ref: "#/components/schemas/LoggedSetDto"
      required: [id, studentId, startTime, loggedSets]
    LoggedSetDto:
      type: object
      properties:
        exerciseId:
          type: string
        setIndex:
          type: integer
        reps:
          type: integer
        weightKg:
          type: number
      required: [exerciseId, setIndex, reps, weightKg]
    CreateSessionRequest:
      type: object
      properties:
        studentId:
          type: string
        planId:
          type: string
        dayId:
          type: string
        startTime:
          type: integer
        endTime:
          type: integer
        exerciseFocus:
          type: string
        loggedSets:
          type: array
          items:
            $ref: "#/components/schemas/LoggedSetDto"
      required: [studentId, startTime, loggedSets]
...