# lib/api-zod/src/generated/types/userDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrainFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserDtoRole } from './userDtoRole';

export interface UserDto {
  id: string;
  name: string;
  email: string;
  role: UserDtoRole;
  onboardingDone?: string;
}