# lib/api-zod/src/generated/types/workoutSessionDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrainFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LoggedSetDto } from './loggedSetDto';

export interface WorkoutSessionDto {
  id: string;
  studentId: string;
  planId?: string;
  dayId?: string;
  startTime: number;
  endTime?: number;
  exerciseFocus?: string;
  loggedSets: LoggedSetDto[];
}